Small change to image preparation
Made a small change to image preparation script. Originally, it assumed you wanted to crop the images in z. This was helpful because I originally only wanted to submit the slices I needed for Quantius (to save on cost). Now, it asks if you want to crop the images in z. This saves you from looking at each stack to make that decision (since it now only asks once).