Finished changes to edit 2D segmentations GUI.
Finished changes to edit 2D segmentations GUI. Biggest changes are that:

(1) The GUI is now a utility (rather than in review_2D_segmentations).
(2) The GUI will better accomodate nucleAIzer segmentations.
(3) You can change the channels displayed, used for calculation, etc in the GUI. This is helpful when the color of the segmentations matches the color of the channel being displayed.