Updated guess_3D_segmentations to run better in batches.
Updated guess_3D_segmentations so that it can optionally accpet an input (which dictates which segmentation method to use). Previously, the code would use a UI to ask the user to pick a segmentation method. The code will still do this, but only if you don't give an input.

This makes it easier to run the code on batches of folders (instead of a single folder each time).