Added a new algorithm for connecting segmentations in 3D.
This new algorithm just assigns all 2D objects to the same 3D object. This is useful for MDCK organoids that have been treated with HGF and are very spindel-y. I was using the connected components algorithm to convert organoid segmentations to 3D, but this does sometimes not properly connect MDCK HGF organoid segmentations.