Fixed small (but important bug) to collecting segmentations and made a few changes to feature calculations.
Fixed a very small (but very important bug) in collect all segmentations steps. The bug was flipping the XY and coordinates for all child structures (for lumens and nuclei when the parent struct is organoid, etc).

Fixed the cell width feature to work even if there are no nuclei.

A few small changes to feature measurement:
- removed parfor loop because solidity feature was crashing with it
- saved the "middle slice" of the organoid so it can be re-used elsewhere
- renamed the radius features (was radius normal, now is radius XY)